20180413
引用DSP库,测试数据信息改用浮点数float变量

+++ BSP_USER/Data_Math/Data_Math.h
@@ -2,9 +2,9 @@
 #define _DATA_MATH_H
 #include "sys.h"
 
-u16 Get_PowerVoltage(void);
-u32 Get_PowerCurrent(void);
-u16 Get_PowerRipple(void);
+float Get_PowerVoltage(void);
+float Get_PowerCurrent(void);
+float Get_PowerRipple(void);
 
 #endif
